Ändrat struktur, lagt till mappar.
