+++ src/Components/AccountSettings/LogOutAlert.js
