+++ src/Components/Pages/LoginPage.js
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
-import Header from "../Header";
-import Login from "../Login";
+import Header from "../LoginAndSignUp/Header";
+import Login from "../LoginAndSignUp/Login";
 
 const LoginPage = ({ setLoggedInUser, loggedInUser }) => {
   return (
