+++ src/Components/Pages/SignUpPage.js
@@ -1,5 +1,5 @@
-import SignUp from "../SignUp";
-import Header from "../Header";
+import SignUp from "../LoginAndSignUp/SignUp";
+import Header from "../LoginAndSignUp/Header";
 
 export const SignUpPage = () => {
   return (
